Use compose value in Relations
