+++ src/Twig/RelatedExtension.php
@@ -9,7 +9,7 @@ use Bolt\Entity\Content;
 use Bolt\Entity\Relation;
 use Bolt\Repository\RelationRepository;
 use Bolt\Storage\Query;
-use Bolt\Utils\Excerpt;
+use Bolt\Utils\ComposeValueHelper;
 use Tightenco\Collect\Support\Collection;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
@@ -122,7 +122,7 @@ class RelatedExtension extends AbstractExtension
         return null;
     }
 
-    public function getRelatedOptions(string $contentTypeSlug, ?string $order = null): Collection
+    public function getRelatedOptions(string $contentTypeSlug, ?string $order = null, string $format = ''): Collection
     {
         $maxAmount = $this->config->get('maximum_listing_select', 1000);
 
@@ -144,12 +144,7 @@ class RelatedExtension extends AbstractExtension
         foreach ($records as $record) {
             $options[] = [
                 'key' => $record->getId(),
-                'value' => sprintf(
-                    '%s (№ %s, %s)',
-                    Excerpt::getExcerpt($record->getExtras()['title'], 50),
-                    $record->getId(),
-                    $record->getStatus()
-                ),
+                'value' => ComposeValueHelper::get($record, $format),
             ];
         }
