+++ templates/content/_relationships.html.twig
@@ -1,7 +1,7 @@
 
 {% for contentType, relation in record.definition.relations %}
 
-    {% set options = related_options(contentType, relation.order|default()) %}
+    {% set options = related_options(contentType, relation.order|default(), relation.format|default()) %}
     {% set value = record|related_values(contentType) %}
 
     <div class="form-group is-normal">
